requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-appconf==1.0.2
django-imagekit==3.3
gunicorn==19.6.0
pilkit==1.1.13
Pillow==3.2.0
six==1.10.0
wheel==0.26.0
whitenoise==3.2
psycopg2==2.6.1